Move Factory And Config Up One Level
