+++ internal/storage/v2/clickhouse/config_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package tracestore
+package clickhouse
 
 import (
 	"testing"
